Similar Words Finder Application/requirements.txt
langchain
streamlit
openai
tiktoken
python-dotenv
faiss-cpu
